# openapi.yaml
openapi: 3.0.3
info:
  title: Polo Sanca Refrigeration Monitoring API
  description: |
    API for monitoring and managing refrigeration equipment across multiple companies and branches.

    ## Authentication
    - **User Authentication**: JWT tokens via `Authorization: Bearer <token>` header
    - **Equipment Authentication**: API keys via `X-API-Key: <key>` header

    ## Versioning
    - Current version: v1
    - Version included in URL path
    - Previous versions supported for 12 months after new release

  version: 1.0.0
  contact:
    name: Polo Sanca API Support
    email: api@polosanca.com
  license:
    name: Proprietary
servers:
  - url: https://api.polosanca.com/v1
    description: Production server
  - url: https://api-staging.polosanca.com/v1
    description: Staging server
...
paths:
...
  /companies:
    get:
      tags:
        - Companies
      summary: List companies
      description: Get list of all companies (Global Admin only)
      operationId: listCompanies
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  companies:
                    type: array
                    items:
                      $ref: '#/components/schemas/Company'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Companies
      summary: Create company
      description: Create a new company (Global Admin only)
      operationId: createCompany
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                contact_email:
                  type: string
                  format: email
                contact_phone:
                  type: string
      responses:
        '201':
          description: Company created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
  /companies/{company_id}:
    get:
      tags:
        - Companies
      summary: Get company details
      operationId: getCompany
      security:
        - bearerAuth: []
      parameters:
        - name: company_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'

    patch:
      tags:
        - Companies
      summary: Update company
      operationId: updateCompany
      security:
        - bearerAuth: []
      parameters:
        - name: company_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                contact_email:
                  type: string
                contact_phone:
                  type: string
      responses:
        '200':
          description: Company updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /equipments:
    get:
      tags:
        - Equipment
      summary: List equipment
      description: Get equipment filtered by user permissions
      operationId: listEquipment
      security:
        - bearerAuth: []
      parameters:
        - name: company_id
          in: query
          description: Filter by company (Global Admin only)
          schema:
            type: string
        - name: branch_id
          in: query
          description: Filter by branch
          schema:
            type: string
        - name: status
          in: query
          description: Filter by status
          schema:
            type: string
            enum: [operational, warning, critical, offline]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  equipments:
                    type: array
                    items:
                      $ref: '#/components/schemas/Equipment'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Equipment
      summary: Register new equipment
      operationId: createEquipment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - serial
                - type
                - branch_id
              properties:
                serial:
                  type: string
                type:
                  type: string
                  enum: [freezer, refrigerator, cold_room]
                branch_id:
                  type: string
                manufacturer:
                  type: string
                model:
                  type: string
                api_key:
                  type: string
      responses:
        '201':
          description: Equipment registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
  /equipments/{equipment_id}:
    get:
      tags:
        - Equipment
      summary: Get equipment details
      operationId: getEquipment
      security:
        - bearerAuth: []
      parameters:
        - name: equipment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'

    patch:
      tags:
        - Equipment
      summary: Update equipment
      operationId: updateEquipment
      security:
        - bearerAuth: []
      parameters:
        - name: equipment_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                manufacturer:
                  type: string
                model:
                  type: string
      responses:
        '200':
          description: Equipment updated successfully

    delete:
      tags:
        - Equipment
      summary: Deactivate equipment
      operationId: deleteEquipment
      security:
        - bearerAuth: []
      parameters:
        - name: equipment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Equipment deactivated successfully
...
  /equipments/{equipment_id}/maintenance-records:
    get:
      tags:
        - Maintenance
      summary: Get maintenance history
      description: Get maintenance records for equipment
      operationId: getMaintenanceRecords
      security:
        - bearerAuth: []
      parameters:
        - name: equipment_id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  equipment_id:
                    type: string
                  records:
                    type: array
                    items:
                      $ref: '#/components/schemas/MaintenanceRecord'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Maintenance
      summary: Log maintenance activity
      operationId: createMaintenanceRecord
      security:
        - bearerAuth: []
      parameters:
        - name: equipment_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - description
                - performed_by
              properties:
                type:
                  type: string
                  example: routine_inspection
                description:
                  type: string
                performed_by:
                  type: string
                notes:
                  type: string
                next_maintenance_date:
                  type: string
                  format: date
      responses:
        '201':
          description: Maintenance record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MaintenanceRecord'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Equipment API key for telemetry submission
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, inactive]
        branches_count:
          type: integer
        equipment_count:
          type: integer
        contact_email:
          type: string
        contact_phone:
          type: string
        created_at:
          type: string
          format: date-time
...
    Equipment:
      type: object
      properties:
        id:
          type: string
        serial:
          type: string
        type:
          type: string
          enum: [freezer, refrigerator, cold_room]
        company_id:
          type: string
        branch_id:
          type: string
        branch_name:
          type: string
        status:
          type: string
          enum: [operational, warning, critical, offline]
        current_data:
          $ref: '#/components/schemas/CurrentData'
        manufacturer:
          type: string
        model:
          type: string
        installed_at:
          type: string
          format: date-time
    CurrentData:
      type: object
      properties:
        temperature:
          type: number
          format: float
        pressure:
          type: number
          format: float
        door:
          type: string
          enum: [open, closed]
        heater:
          type: string
          enum: [on, off]
        compressor:
          type: string
          enum: [on, off]
        fan:
          type: string
          enum: [on, off]
        last_update:
          type: string
          format: date-time
...
    MaintenanceRecord:
      type: object
      properties:
        id:
          type: string
        equipment_id:
          type: string
        type:
          type: string
        description:
          type: string
        performed_by:
          type: string
        performed_at:
          type: string
          format: date-time
        notes:
          type: string
        next_maintenance_date:
          type: string
          format: date
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
        limit:
          type: integer
          description: Items per page
        total:
          type: integer
          description: Total number of items
        total_pages:
          type: integer
          description: Total number of pages
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
...